backend/requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-jose[cryptography]>=3.3.0
python-multipart>=0.0.6
pydantic>=2.0.0
requests>=2.31.0
